# schemas/openapi/lab.yaml
openapi: 3.0.3
info:
  title: AtlasLab HTTP API
  version: atlaslab.v1
  description: |
    Operator-console Lab control and evidence API.
    AtlasLab is the supported Agent Host and evaluator. It is not merchant authority.
    Deterministic runs do not require OpenRouter. Razorpay is Test Mode only.
servers:
  - url: http://127.0.0.1:8090
paths:
...
  /lab/v1/runs:
    post:
      operationId: startRun
      tags: [lab]
      security:
        - LabBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartRunRequest"
      responses:
        "201":
          description: Run created and executed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEnvelope"
        "400":
          description: Wrong variant or validation error
        "401":
          description: Missing bearer
        "403":
          description: Forbidden
        "409":
          description: Model runs unavailable
  /lab/v1/runs/{run_id}:
    get:
      operationId: getRun
      tags: [lab]
      parameters:
        - $ref: "#/components/parameters/RunId"
      responses:
        "200":
          description: Run summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RunEnvelope"
...
  /lab/v1/pairs:
    post:
      operationId: startPair
      tags: [lab]
      security:
        - LabBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartPairRequest"
      responses:
        "201":
          description: Pair created; assignment frozen before arm execution (ID-602)
...
  /lab/v1/pairs/{pair_id}:
    get:
      operationId: getPair
      tags: [lab]
      parameters:
        - in: path
          name: pair_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Pair eligibility, states, deltas, guardrails
  /lab/v1/reports/{report_id}:
    get:
      operationId: getReport
      tags: [lab]
      parameters:
        - in: path
          name: report_id
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generated JSON/CSV report labeled Razorpay Test Mode
components:
  securitySchemes:
    LabBearer:
      type: http
      scheme: bearer
  parameters:
    RunId:
      in: path
      name: run_id
      required: true
      schema:
        type: string
  schemas:
    RunType:
      type: string
      enum: [DETERMINISTIC_SCENARIO, BENCHMARK_MODEL, CUSTOM_MISSION]
    EvidenceEligibility:
      type: string
      enum: [CONTRACT_EVIDENCE_ONLY, BENCHMARK_ELIGIBLE, BENCHMARK_INELIGIBLE, EXPLORATORY]
...
    StartRunRequest:
      type: object
      required: [run_type]
      properties:
        run_type:
          $ref: "#/components/schemas/RunType"
        scenario_id:
          type: string
        fixture_snapshot_id:
          type: string
        payment_simulation:
          type: string
        model_id:
          type: string
          description: Required for BENCHMARK_MODEL and CUSTOM_MISSION. Forbidden on DETERMINISTIC_SCENARIO.
        custom_user_input:
          type: string
          maxLength: 4000
        custom_input_digest:
          type: string
        consent_policy:
          type: object
        permitted_actions:
          type: array
          items:
            type: string
        arm:
          type: string
          enum: [CONTROL, TREATMENT]
        pairing_key:
          type: string
    RunSummary:
      type: object
      required: [run_id, run_type, evidence_eligibility, state, configuration_digest]
      properties:
        run_id:
          type: string
        run_type:
          $ref: "#/components/schemas/RunType"
        evidence_eligibility:
          $ref: "#/components/schemas/EvidenceEligibility"
        state:
          type: string
        configuration_digest:
          type: string
        fixture_digest:
          type: string
          nullable: true
        requested_model_id:
          type: string
          nullable: true
        returned_model_id:
          type: string
          nullable: true
        arm:
          type: string
          nullable: true
        pair_id:
          type: string
          nullable: true
        terminal_reason:
          type: string
          nullable: true
    RunEnvelope:
      type: object
      properties:
        request_id:
          type: string
        occurred_at:
          type: string
          format: date-time
        run:
          $ref: "#/components/schemas/RunSummary"
...
    StartPairRequest:
      type: object
      required: [pairing_key]
      properties:
        pairing_key:
          type: string
        scenario_id:
          type: string
        model_id:
          type: string
        fixture_snapshot_id:
          type: string
        payment_simulation:
          type: string
        control_run_id:
          type: string
          description: Legacy diagnostic only. Pairing finished runs cannot enter denominators (ID-602).
        treatment_run_id:
          type: string